Created array to store all posts, added comments
I learned a lot about using the array. We were introduced to the spread operator and the map() function. It turned out the teacher did have an issue with refreshing. I learned a third method to combat the problem - change the form into a div because forms can cause default refreshing.